// clustrum-portal/src/shared/api/navigation/navigation-list/types/navigation-list-request-params.ts
import { Order } from '../../../../lib/constants/order';
import { Places } from '../../../../lib/constants/places';

export interface NavigationListRequestParams {
  orderBy?: Order;
  page: number | null;
  pageSize: number;
  path: string;
  place: Places | null;
  placeParameters?: PlaceParameters; // TODO: поле, возможно, не нужно, удалить после уточнения api
}

interface PlaceParameters {
  displayParentFolder: boolean;
  filters: boolean;
  icon: Icon;
  iconClassName: string;
  pagination: boolean;
  place: string;
  sort: boolean;
  text: string;
}

interface Icon {
  id: string;
  url: string;
  viewBox: string;
}